interview-service-ai/VideoAudioSeperatorService/requirements.txt
requests==2.32.5
moviepy==1.0.3
cloudinary==1.44.1
python-dotenv==1.1.1
kafka-python==2.2.15